prism_tier2/requirements.txt
# PRISM - Tier 2 Production Dependencies
# Install with: pip install -r requirements.txt

# Core
pyyaml>=6.0
flask>=3.0

# Production web server (replaces Flask dev server)
gunicorn>=21.0
gevent>=23.0

# Task queue
celery>=5.3
redis>=5.0

# Database (audit log, review queue, job state)
psycopg2-binary>=2.9
sqlalchemy>=2.0

# File watching
watchdog>=4.0

# Monitoring (optional - Celery task dashboard at :5555)
flower>=2.0

# Format parsers (optional but recommended)
python-evtx>=0.7.4      # Windows .evtx binary log parsing
